IS: Move aktivitetskrav folder into personstatus folder structure
