--- src/main/kotlin/no/nav/syfo/aktivitetskravvurdering/domain/Aktivitetskrav.kt
@@ -1,15 +1,0 @@
-package no.nav.syfo.aktivitetskravvurdering.domain
-
-enum class AktivitetskravStatus {
-    NY,
-    NY_VURDERING,
-    AVVENT,
-    UNNTAK,
-    OPPFYLT,
-    FORHANDSVARSEL,
-    AUTOMATISK_OPPFYLT,
-    STANS,
-    IKKE_OPPFYLT,
-    IKKE_AKTUELL,
-    LUKKET,
-}
