+++ src/main/kotlin/no/nav/syfo/personstatus/infrastructure/kafka/aktivitetskrav/AktivitetskravVurderingConsumer.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.aktivitetskravvurdering.kafka
+package no.nav.syfo.personstatus.infrastructure.kafka.aktivitetskrav
 
 import no.nav.syfo.personstatus.application.PersonoversiktStatusService
 import no.nav.syfo.personstatus.domain.PersonIdent
@@ -7,6 +7,12 @@ import org.apache.kafka.clients.consumer.ConsumerRecords
 import org.apache.kafka.clients.consumer.KafkaConsumer
 import org.slf4j.LoggerFactory
 import java.time.Duration
+import java.time.LocalDate
+import java.time.OffsetDateTime
+import kotlin.collections.forEach
+import kotlin.collections.isNotEmpty
+import kotlin.collections.partition
+import kotlin.jvm.java
 
 class AktivitetskravVurderingConsumer(
     private val personoversiktStatusService: PersonoversiktStatusService,
@@ -42,3 +48,15 @@ class AktivitetskravVurderingConsumer(
         private val log = LoggerFactory.getLogger(AktivitetskravVurderingRecord::class.java)
     }
 }
+
+data class AktivitetskravVurderingRecord(
+    val uuid: String,
+    val personIdent: String,
+    val createdAt: OffsetDateTime,
+    val status: String,
+    val isFinal: Boolean,
+    val stoppunktAt: LocalDate,
+    val beskrivelse: String?,
+    val sistVurdert: OffsetDateTime?,
+    val frist: LocalDate?,
+)
